requirements.txt
click
Pillow
pyfiglet
colorama
pdf2image
pdfplumber
python-docx
